Equally smooth zoom
Previously if picture in zoom 1:1 zoom larger it will zooms more times faster than if zoom it from from zoom 1:1 to smaller size. Now zoom speed is equal.

+++ src/cpp_h/ImageViewer/pixmapimage.h
@@ -13,7 +13,7 @@ public:
     Q_INVOKABLE void setImage(QObject *pixContainer);
     Q_INVOKABLE void setViewerHeight(const int dimension);
     Q_INVOKABLE void setViewerWidth(const int dimension);
-    Q_INVOKABLE void setCanvasScale(const int number);
+    Q_INVOKABLE void setCanvasScale(const qreal number);
 
 protected:
     virtual void paint(QPainter *painter) Q_DECL_OVERRIDE;
